# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateTableRequest:
    properties:
      table_name:
        maxLength: 50
        type: string
    required:
    - table_name
    type: object
...
  dto.UpdateTableRequest:
    properties:
      table_name:
        maxLength: 50
        type: string
    required:
    - table_name
    type: object
...
info:
  contact: {}
  description: Backend API Service untuk platform SaaS POS & Self-Order UMKM kuliner.
  title: saas_gangsta Backend API
  version: "1.0"
paths:
...
  /dining-tables:
    get:
      description: Mengambil daftar meja per tenant
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Tables
      tags:
      - Dining Table Management
    post:
      consumes:
      - application/json
      description: Mendaftarkan meja baru
      parameters:
      - description: Payload Create Table
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTableRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Table
      tags:
      - Dining Table Management
  /dining-tables/{id}:
    delete:
      description: Menghapus meja (soft delete)
      parameters:
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Soft Delete Table
      tags:
      - Dining Table Management
    get:
      description: Mengambil detail satu meja
      parameters:
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Detail Table
      tags:
      - Dining Table Management
    put:
      consumes:
      - application/json
      description: Memperbarui nama meja
      parameters:
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload Update Table
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Table
      tags:
      - Dining Table Management
  /dining-tables/{id}/status:
    get:
      description: Mengecek apakah meja sedang kosong atau occupied (berdasarkan transaksi
        pesanan)
      parameters:
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cek Status Meja
      tags:
      - Dining Table Management
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"